% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{shinyDCQC}
\alias{shinyDCQC}
\title{Run the DCQC reviewing app}
\usage{
shinyDCQC()
}
\value{
A PDF of the review
}
\description{
The SORTEE DCQC reviewing app allows data editors to review papers accroding to the SORTEE guidelines.
}
